Move strategy tests from accelerators to strategies directory
