+++ tests/strategies/test_ddp.py
@@ -23,9 +23,9 @@ from torch.nn.parallel.distributed import DistributedDataParallel
 import pytorch_lightning as pl
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import Callback
-from tests.accelerators import ddp_model
 from tests.helpers.boring_model import BoringModel
 from tests.helpers.runif import RunIf
+from tests.strategies import ddp_model
 from tests.utilities.distributed import call_training_script
 
 CLI_ARGS = "--max_epochs 1 --gpus 2 --strategy ddp"
